refactor: rename fs utility to create sandboxed fs
